Restore bootstrap
Okay, I understand the desire to do this, but we can't.  Not yet.  Open
an issue.

The behavior we need is that we don't have to manually muck around with
the bootstrap files when we organize the repository.  We can actually
generate them from a (working) atbuild, check them into source control,
and then bootstrap can use them without any atbuild installed.

The catch is, they need to talk to the .atllbuild directory, not some
other .bootstrap directory.

To get the latter behavior, open an issue and I can add support for this
in atllbuild.

This commit restores the behavior where we can generate the bootstrap
files from a working atbuild again, so we don't have to edit them by
hand.