Support for importing remote tasks from another file
We have to be careful about how these paths work; because when you
import a new file it is as if your working directory changes.  To
support this, we add a new `importedPath` property to Task that contains
the path where the task is imported (whether local or remote).  This
property is documented to include a trailing slash to avoid various
problems involving relative paths.

We also fix #20 by moving to `collectSources` implemented in atpkg
instead of atllbuild.