Add atbuild "user paths".
This patch adds a "user path" which is global to an entire buildchain
and allows tasks to progressively build up contents that are somewhat
arbitrary.

In particular, this is useful for AnarchyDispatch.  If we have
a/b/AnarchyDispatch, then AnarchyDispatch can install itself to the
ATBUILD_USER_PATH and be visible to both a and b.  Then we can inject
includeWithUser: ["AnarchyDispatch"] in an overlay and everything "just
works".