Allow variable expansion for shell tool
This allows the execution of a shell tool with the "collected sources"
of another task.

The intent here is to allow the creation of arbitrary tasks that accept
source files as arguments. xcode-emit is an obvious candidate, and
would allow it to break its dependence on atpkg, which it only uses for
this single feature.

There are other obvious programs that we might want to call with "all
sources" (e.g. preprocessors, etc.) and they may not want to take a
dependency on atpkg either.

See https://github.com/AnarchyTools/atpkg/pull/23