Add executable-name option to atllbuild
We add an executable-name option to atllbuild, allowing the use of "non-module-safe" names for executables.  This includes e.g. hyphens, which are a legal executable name but not a legal module-name.

Resolve #27.  This resolution was chosen (over name/module-name) because the module-name is used in several places (such as Frameworks for example) and the executable case seems to be the odd one out at present.